+++ src/Organisms/Development/Development.stories.tsx
@@ -3,11 +3,11 @@ import { storiesOf } from '@storybook/react';
 import MD from 'react-markdown';
 
 import docs from './Development.md';
-import { Development, Typography } from '../..';
+import { Development, Typography } from '../../index';
 import { Display } from '../../common/Display/index';
 import { TYPOGRAPHY_TYPES } from '../../Atoms/Typography/Typography';
 
-storiesOf('Molecules | Development', module)
+storiesOf('Organisms | Development', module)
   .add('Documentation', () => (
     <Typography>
       <MD source={docs} />
